Moved code that converted lightmap texels from unsigned int values to floating point values, in the OXR version, to the moment they are copied into the staging buffer - this way, the conversion is done only when the rendering engine really needs it, thus improving performance a bit. Minor optimizations applied to the code that generates the lightmaps themselves in the core engine.

+++ OpenXR-SDK/src/SlipNFrag-XR/app/src/main/cpp/LoadedLightmap.h
@@ -6,6 +6,6 @@ struct LoadedLightmap
 {
 	Lightmap* lightmap;
 	VkDeviceSize size;
-	float* source;
+	unsigned* source;
 	LoadedLightmap* next;
 };
